Removed memoization hooks in team analysis #671
