+++ src/ui/team/analysis/calc/hook/main.ts
@@ -1,5 +1,3 @@
-import React from 'react';
-
 import {MealCoverage} from '@/types/game/cooking';
 import {ProducingRate} from '@/types/game/producing/rate';
 import {getTeamCompCalcResult} from '@/ui/team/analysis/calc/comp';
@@ -14,28 +12,20 @@ import {getTotalOfGroupedProducingRate} from '@/utils/game/producing/rateReducer
 
 export const useTeamProducingStats = (opts: UseTeamProducingStatsOpts): TeamProducingStats => {
   const {
-    setup,
-    bundle,
     currentTeam,
     cookingSettings,
   } = opts;
   const {
     snorlaxFavorite,
     analysisPeriod,
-    members,
   } = currentTeam;
 
-  const deps: React.DependencyList = [snorlaxFavorite, analysisPeriod, members, setup, bundle];
-
-  const compsStats = React.useMemo(
-    () => getTeamCompCalcResult({
-      period: analysisPeriod,
-      state: stateOfRateToShow,
-      snorlaxFavorite,
-      ...opts,
-    }),
-    deps,
-  );
+  const compsStats = getTeamCompCalcResult({
+    period: analysisPeriod,
+    state: stateOfRateToShow,
+    snorlaxFavorite,
+    ...opts,
+  });
   const {
     bySlot,
     grouped,
@@ -45,20 +35,19 @@ export const useTeamProducingStats = (opts: UseTeamProducingStatsOpts): TeamProd
     period: analysisPeriod,
     bySlot,
     state: stateOfRateToShow,
-    deps,
   });
 
-  const overall: ProducingRate = React.useMemo(() => ({
+  const overall: ProducingRate = {
     period: analysisPeriod,
     energy: getTotalOfGroupedProducingRate({rate: total, key: 'energy'}),
     quantity: getTotalOfGroupedProducingRate({rate: total, key: 'quantity'}),
-  }), deps);
+  };
 
-  const mealCoverage: MealCoverage = React.useMemo(() => getMealCoverage({
+  const mealCoverage: MealCoverage = getMealCoverage({
     meals: cookingSettings.targetMeals,
     ingredientProduction: toIngredientProductionCounterFromGroupedRate(grouped.ingredient),
     period: analysisPeriod,
-  }), deps);
+  });
 
   return {
     total,
